refactor: split tzadikim.ts into per-story files (#485)
Extract each tzaddik story into its own file under data/stories/,
introduce a shared TzaddikStory type in data/types.ts, and reduce
tzadikim.ts to a thin barrel that assembles TZADIKIM_STORIES array
in the original order. No consumer import paths changed.
