+++ gamesformykids/components/game/quiz/games/tzadikim/data/stories/baalShemTov.ts
@@ -1,0 +1,39 @@
+import type { TzaddikStory } from '../types';
+
+export const BAAL_SHEM_TOV_STORY: TzaddikStory = {
+  id: 'baal-shem-tov',
+  name: 'הבעל שם טוב',
+  years: 'תנ"ח – ק"כ (1698–1760)',
+  emoji: '✨',
+  color: 'from-blue-500 to-indigo-600',
+  bgGradient: 'from-blue-50 to-indigo-100',
+  shortTitle: 'מייסד החסידות',
+  story: `רבי ישראל בן אליעזר, המכונה "הבעל שם טוב", גדל כיתום קטן בעיירה קטנה באוקראינה. מגיל צעיר היה אוהב ללכת ביערות ולהתפלל לה׳ בשמחה גדולה.
+
+הבעל שם טוב לימד שאפשר לעבוד את ה׳ בשמחה ואהבה, ולא רק בלימוד ודאגה. הוא אמר שגם פשוט אנשים שאינם יכולים ללמוד הרבה, יכולים להתקרב לה׳ בלב שלם.
+
+פעם אחת בא אליו יהודי פשוט שאמר: "רבי, אני לא יודע ללמוד, ואני יודע רק לשרוק בפה". אמר לו הבעל שם טוב: "שרוק לפני ה׳ בכל כוחך!" שרק האיש חזק ומתוק, ואמר הבעל שם טוב: "השריקה הזו עלתה לשמים ופתחה שערים שלא נפתחו שנים!"`,
+  lesson: 'כל אדם יכול להתקרב לה׳ בדרכו שלו, בשמחה ובאהבה.',
+  questions: [
+    {
+      question: 'מה השם האמיתי של הבעל שם טוב?',
+      answers: ['רבי עקיבא', 'רבי ישראל בן אליעזר', 'רבי שמעון', 'רבי יוסי'],
+      correctIndex: 1,
+    },
+    {
+      question: 'מה לימד הבעל שם טוב?',
+      answers: [
+        'שצריך ללמוד הרבה',
+        'שאפשר לעבוד ה׳ בשמחה ואהבה',
+        'שצריך להיות עצוב',
+        'שרק חכמים יכולים להתפלל',
+      ],
+      correctIndex: 1,
+    },
+    {
+      question: 'מה עשה האיש הפשוט שבא אל הבעל שם טוב?',
+      answers: ['שר שיר', 'ריקד', 'שרק בפה', 'קרא מהסידור'],
+      correctIndex: 2,
+    },
+  ],
+};
