+++ gamesformykids/components/game/quiz/games/tzadikim/data/stories/chofetzChaim.ts
@@ -1,0 +1,50 @@
+import type { TzaddikStory } from '../types';
+
+export const CHOFETZ_CHAIM_STORY: TzaddikStory = {
+  id: 'chofetz-chaim',
+  name: 'החפץ חיים',
+  years: 'תקנ"ח – תרצ"ג (1838–1933)',
+  emoji: '🕊️',
+  color: 'from-green-500 to-teal-600',
+  bgGradient: 'from-green-50 to-teal-100',
+  shortTitle: 'שומר הלשון',
+  story: `רבי ישראל מאיר הכהן מראדין, המכונה "החפץ חיים" על שם ספרו, היה אחד הצדיקים הגדולים של הדור האחרון.
+
+פעם בא אדם לרבי ושאל: "רבי, חטאתי בלשון הרע – מה אעשה?" ענה הרבי: "לך לשוק, קח כר נוצות, ועלה לגג הגבוה ביותר, ושחרר את כל הנוצות לרוח."
+
+עשה האיש כך. חזר אל הרבי ואמר: "עשיתי!" ואמר לו הרבי: "עכשיו לך ואסוף את כל הנוצות!"
+
+"אבל רבי, זה בלתי אפשרי! הרוח פיזרה אותן לכל עבר!"
+
+"כך הם דברי לשון הרע," אמר הרב בעצב. "ברגע שיוצאים מהפה – אי אפשר לאספם בחזרה."
+
+החפץ חיים לימד אותנו לשמור על לשוננו ולדבר רק דברים טובים.`,
+  lesson: 'דברים שאומרים יכולים לפגוע כמו חצים. חשוב לחשוב לפני שמדברים.',
+  questions: [
+    {
+      question: 'על שם מה נקרא "החפץ חיים"?',
+      answers: ['על שם עירו', 'על שם ספרו', 'על שם אביו', 'על שם רבו'],
+      correctIndex: 1,
+    },
+    {
+      question: 'מה ביקש הרב מהאדם לעשות?',
+      answers: [
+        'לבנות בית כנסת',
+        'לצום שלושה ימים',
+        'לשחרר נוצות מהגג',
+        'לתת צדקה',
+      ],
+      correctIndex: 2,
+    },
+    {
+      question: 'מה לימד החפץ חיים בסיפור הנוצות?',
+      answers: [
+        'שהרוח חזקה',
+        'שדברי לשון הרע אי אפשר לקחת בחזרה',
+        'שצריך לנקות את הבית',
+        'שגגות גבוהים מסוכנים',
+      ],
+      correctIndex: 1,
+    },
+  ],
+};
