+++ gamesformykids/components/game/quiz/games/tzadikim/data/stories/menachamMendelMiVitebsk.ts
@@ -1,0 +1,48 @@
+import type { TzaddikStory } from '../types';
+
+export const MENACHEM_MENDEL_MI_VITEBSK_STORY: TzaddikStory = {
+  id: 'the-alter-of-kelm',
+  name: "הרב מנחם מנדל מוויטבסק",
+  years: 'תק"ל – תקנ"ח (1730–1788)',
+  emoji: '🌿',
+  color: 'from-lime-500 to-green-600',
+  bgGradient: 'from-lime-50 to-green-100',
+  shortTitle: 'אהבת ארץ ישראל',
+  story: `רבי מנחם מנדל מוויטבסק היה תלמיד גדול של המגיד ממזריץ', ממשיכי הבעל שם טוב.
+
+מתוך אהבה עזה לארץ ישראל, עלה רבי מנחם מנדל לארץ ישראל עם שלוש מאות חסידים בשנת תקל"ז. זה היה מסע קשה מאוד – חודשים של הפלגה בים ומסע יבשה.
+
+הם נקלטו בצפת ואחר כך עברו לטבריה. החיים היו קשים מאוד – עוני, מחלות, וקשיים רבים. חסידיו שאלוהו פעמים רבות: "רבי, למה לא נחזור?"
+
+ענה רבם: "אני ישן על אדמת ישראל, ואפילו שעה אחת כאן שווה יותר מכל שנות החיים שם!"
+
+גם מרחוק המשיך לכתוב אגרות עידוד לחסידיו שנשארו בגלות, ועודדם לשמור אמונה ולחכות לגאולה.`,
+  lesson: 'ארץ ישראל היא ארץ מיוחדת, ואהבתה שווה כל מחיר.',
+  questions: [
+    {
+      question: 'מה עשה רבי מנחם מנדל מתוך אהבה לארץ ישראל?',
+      answers: [
+        'כתב ספר על הארץ',
+        'שלח כסף לעניים',
+        'עלה לארץ ישראל עם שלוש מאות חסידים',
+        'לימד ילדים על הארץ',
+      ],
+      correctIndex: 2,
+    },
+    {
+      question: 'באיזו עיר התיישבו החסידים בתחילה?',
+      answers: ['ירושלים', 'חברון', 'צפת', 'באר שבע'],
+      correctIndex: 2,
+    },
+    {
+      question: 'מה אמר הרב כשרצו לחזור?',
+      answers: [
+        '"אז נחזור"',
+        '"שעה אחת בארץ ישראל שווה יותר מכל שנות החיים שם"',
+        '"זה קשה מדי"',
+        '"צדקתם, נחזור"',
+      ],
+      correctIndex: 1,
+    },
+  ],
+};
