+++ gamesformykids/components/game/quiz/games/tzadikim/data/stories/rabbiAkiva.ts
@@ -1,0 +1,41 @@
+import type { TzaddikStory } from '../types';
+
+export const RABBI_AKIVA_STORY: TzaddikStory = {
+  id: 'rabbi-akiva',
+  name: 'רבי עקיבא',
+  years: "כ\"ז לפנה\"ס – ק\"לה (±40–135 לספירה)",
+  emoji: '📖',
+  color: 'from-amber-500 to-orange-600',
+  bgGradient: 'from-amber-50 to-orange-100',
+  shortTitle: 'מאות לגדול בתורה',
+  story: `רבי עקיבא היה רועה צאן פשוט עד גיל ארבעים ולא ידע קרוא וכתוב. אשתו רחל האמינה בו ועודדה אותו ללמוד תורה.
+
+פעם אחת ראה רבי עקיבא סלע שנחצב על ידי מים. שאל את עצמו: "אם מים רכים יכולים לחצוב סלע קשה, בוודאי דברי תורה יכולים לחצוב את לבי!"
+
+עזב את עדרו והחל ללמוד אלף-בית עם ילדים קטנים. למרות שכולם צחקו עליו, הוא לא הפסיק. בסוף חייו היה לגדול הדור, ועשרים וארבעה אלף תלמידים ישבו לרגליו!
+
+רבי עקיבא לימד: "ואהבת לרעך כמוך – זה כלל גדול בתורה."`,
+  lesson: 'אף פעם לא מאוחר ללמוד. עם נחישות ואהבה אפשר להגיע לכל מקום.',
+  questions: [
+    {
+      question: 'מה עשה רבי עקיבא לפני שהתחיל ללמוד תורה?',
+      answers: ['לימד ילדים', 'היה רועה צאן', 'בנה בתים', 'מכר בשוק'],
+      correctIndex: 1,
+    },
+    {
+      question: 'מה עורר את רבי עקיבא להתחיל ללמוד?',
+      answers: [
+        'חלום שחלם',
+        'מים שחצבו סלע',
+        'ספר שמצא',
+        'חכם שפגש',
+      ],
+      correctIndex: 1,
+    },
+    {
+      question: 'כמה תלמידים היו לרבי עקיבא?',
+      answers: ['אלף', 'עשרה אלפים', 'עשרים וארבעה אלף', 'מאה'],
+      correctIndex: 2,
+    },
+  ],
+};
