+++ gamesformykids/components/game/quiz/games/tzadikim/data/stories/rabbiMeirBaalHanes.ts
@@ -1,0 +1,55 @@
+import type { TzaddikStory } from '../types';
+
+export const RABBI_MEIR_BAAL_HANES_STORY: TzaddikStory = {
+  id: 'rabbi-meir-baal-hanes',
+  name: "רבי מאיר בעל הנס",
+  years: 'המאה השנייה לספירה',
+  emoji: '🕯️',
+  color: 'from-violet-500 to-purple-600',
+  bgGradient: 'from-violet-50 to-purple-100',
+  shortTitle: 'עושה ניסים',
+  story: `רבי מאיר היה אחד מגדולי תנאי המשנה ותלמיד נאמן של רבי עקיבא. הוא נקרא "בעל הנס" בגלל הניסים שנעשו בזכותו.
+
+אשתו ברוריה הייתה אישה חכמה מאוד. פעם בשבת מתו שני בניהם פתאום. ברוריה לא אמרה לרבי מאיר עד מוצאי שבת, כדי לא לצערו בשבת.
+
+כשסיים שבת, אמרה לו ברוריה: "רבי, שאלה יש לי. מישהו נתן לי לשמור שני תכשיטים יקרים, ועכשיו בא לקחתם חזרה – האם להחזיר?"
+
+"כן, ודאי שמחזירים פיקדון!" ענה הרב.
+
+לקחה אותו ברוריה אל חדר שני ילדיהם ואמרה: "ה׳ נתן וה׳ לקח – יהי שם ה׳ מבורך."
+
+רבי מאיר קיבל את דין השמיים בגבורה ובאמונה שלמה.`,
+  lesson: 'אפשר לקבל גם דברים קשים בחיים עם אמונה ורוגע.',
+  questions: [
+    {
+      question: 'מדוע נקרא "בעל הנס"?',
+      answers: [
+        'כי גר בנס',
+        'כי ניסים נעשו בזכותו',
+        'כי אהב לספר סיפורים',
+        'כי היה מלמד בניסן',
+      ],
+      correctIndex: 1,
+    },
+    {
+      question: 'מה עשתה ברוריה עם ידיעת מות ילדיה?',
+      answers: [
+        'שלחה מיד שליח לבעלה',
+        'בכתה כל השבת',
+        'חיכתה עד מוצאי שבת',
+        'ברחה מהבית',
+      ],
+      correctIndex: 2,
+    },
+    {
+      question: 'מה הייתה תשובת ברוריה לבעלה?',
+      answers: [
+        '"אנחנו עניים"',
+        '"ה׳ נתן וה׳ לקח – יהי שם ה׳ מבורך"',
+        '"הייתה תאונה"',
+        '"הם היו חולים"',
+      ],
+      correctIndex: 1,
+    },
+  ],
+};
