+++ gamesformykids/components/game/quiz/games/tzadikim/data/stories/rabbiShimonBarYochai.ts
@@ -1,0 +1,38 @@
+import type { TzaddikStory } from '../types';
+
+export const RABBI_SHIMON_BAR_YOCHAI_STORY: TzaddikStory = {
+  id: 'rabbi-shimon-bar-yochai',
+  name: 'רבי שמעון בר יוחאי',
+  years: 'המאה השנייה לספירה',
+  emoji: '🔥',
+  color: 'from-red-500 to-rose-600',
+  bgGradient: 'from-red-50 to-rose-100',
+  shortTitle: 'מגלה סודות התורה',
+  story: `רבי שמעון בר יוחאי, המכונה "רשב"י", היה תלמיד של רבי עקיבא ואחד החכמים הגדולים בתלמוד.
+
+כשגזרו הרומאים גזרות קשות על לימוד התורה, ברח רשב"י עם בנו רבי אלעזר ונחבאו במערה שנים-עשר שנה! בזמן הזה למדו תורה כל היום וכל הלילה.
+
+ניס להם נס: צמח עץ חרוב ומזרם מים – כדי שיאכלו וישתו ויוכלו להמשיך ללמוד. בגדיהם הסתירו אותם כשלא למדו, ולבשו אותם רק כשלמדו.
+
+בזכות לימוד התורה הרצוף, גילה רשב"י סודות עמוקים מאוד. ספר הזוהר, הספר הקדוש של הקבלה, מיוחס לו.
+
+בל"ג בעומר, יום פטירתו, ציווה רשב"י לחגוג ולשמוח!`,
+  lesson: 'לימוד התורה הוא אור גדול שמאיר בכל מקום, גם בחשכה.',
+  questions: [
+    {
+      question: 'מי היה רבו של רבי שמעון בר יוחאי?',
+      answers: ['הבעל שם טוב', 'החפץ חיים', 'רבי עקיבא', 'ר׳ יהודה הנשיא'],
+      correctIndex: 2,
+    },
+    {
+      question: 'כמה זמן התחבאו רשב"י ובנו במערה?',
+      answers: ['שנה אחת', 'שלוש שנים', 'שבע שנים', 'שתים-עשרה שנה'],
+      correctIndex: 3,
+    },
+    {
+      question: 'איזה יום חוגגים לזכר רשב"י?',
+      answers: ['ר"ה', 'ל"ג בעומר', 'חנוכה', 'שבועות'],
+      correctIndex: 1,
+    },
+  ],
+};
