+++ gamesformykids/components/game/quiz/games/tzadikim/data/tzadikim.ts
@@ -1,277 +1,17 @@
-import type { QuizQuestion } from '@/lib/types';
-export type { QuizQuestion };
-
-export interface TzaddikStory {
-  id: string;
-  name: string;
-  years: string;
-  emoji: string;
-  color: string;
-  bgGradient: string;
-  shortTitle: string;
-  story: string;
-  lesson: string;
-  questions: QuizQuestion[];
-}
+export type { TzaddikStory, QuizQuestion } from './types';
+import type { TzaddikStory } from './types';
+import { BAAL_SHEM_TOV_STORY } from './stories/baalShemTov';
+import { RABBI_AKIVA_STORY } from './stories/rabbiAkiva';
+import { CHOFETZ_CHAIM_STORY } from './stories/chofetzChaim';
+import { RABBI_SHIMON_BAR_YOCHAI_STORY } from './stories/rabbiShimonBarYochai';
+import { RABBI_MEIR_BAAL_HANES_STORY } from './stories/rabbiMeirBaalHanes';
+import { MENACHEM_MENDEL_MI_VITEBSK_STORY } from './stories/menachamMendelMiVitebsk';
 
 export const TZADIKIM_STORIES: TzaddikStory[] = [
-  {
-    id: 'baal-shem-tov',
-    name: 'הבעל שם טוב',
-    years: 'תנ"ח – ק"כ (1698–1760)',
-    emoji: '✨',
-    color: 'from-blue-500 to-indigo-600',
-    bgGradient: 'from-blue-50 to-indigo-100',
-    shortTitle: 'מייסד החסידות',
-    story: `רבי ישראל בן אליעזר, המכונה "הבעל שם טוב", גדל כיתום קטן בעיירה קטנה באוקראינה. מגיל צעיר היה אוהב ללכת ביערות ולהתפלל לה׳ בשמחה גדולה.
-
-הבעל שם טוב לימד שאפשר לעבוד את ה׳ בשמחה ואהבה, ולא רק בלימוד ודאגה. הוא אמר שגם פשוט אנשים שאינם יכולים ללמוד הרבה, יכולים להתקרב לה׳ בלב שלם.
-
-פעם אחת בא אליו יהודי פשוט שאמר: "רבי, אני לא יודע ללמוד, ואני יודע רק לשרוק בפה". אמר לו הבעל שם טוב: "שרוק לפני ה׳ בכל כוחך!" שרק האיש חזק ומתוק, ואמר הבעל שם טוב: "השריקה הזו עלתה לשמים ופתחה שערים שלא נפתחו שנים!"`,
-    lesson: 'כל אדם יכול להתקרב לה׳ בדרכו שלו, בשמחה ובאהבה.',
-    questions: [
-      {
-        question: 'מה השם האמיתי של הבעל שם טוב?',
-        answers: ['רבי עקיבא', 'רבי ישראל בן אליעזר', 'רבי שמעון', 'רבי יוסי'],
-        correctIndex: 1,
-      },
-      {
-        question: 'מה לימד הבעל שם טוב?',
-        answers: [
-          'שצריך ללמוד הרבה',
-          'שאפשר לעבוד ה׳ בשמחה ואהבה',
-          'שצריך להיות עצוב',
-          'שרק חכמים יכולים להתפלל',
-        ],
-        correctIndex: 1,
-      },
-      {
-        question: 'מה עשה האיש הפשוט שבא אל הבעל שם טוב?',
-        answers: ['שר שיר', 'ריקד', 'שרק בפה', 'קרא מהסידור'],
-        correctIndex: 2,
-      },
-    ],
-  },
-  {
-    id: 'rabbi-akiva',
-    name: 'רבי עקיבא',
-    years: "כ\"ז לפנה\"ס – ק\"לה (±40–135 לספירה)",
-    emoji: '📖',
-    color: 'from-amber-500 to-orange-600',
-    bgGradient: 'from-amber-50 to-orange-100',
-    shortTitle: 'מאות לגדול בתורה',
-    story: `רבי עקיבא היה רועה צאן פשוט עד גיל ארבעים ולא ידע קרוא וכתוב. אשתו רחל האמינה בו ועודדה אותו ללמוד תורה.
-
-פעם אחת ראה רבי עקיבא סלע שנחצב על ידי מים. שאל את עצמו: "אם מים רכים יכולים לחצוב סלע קשה, בוודאי דברי תורה יכולים לחצוב את לבי!"
-
-עזב את עדרו והחל ללמוד אלף-בית עם ילדים קטנים. למרות שכולם צחקו עליו, הוא לא הפסיק. בסוף חייו היה לגדול הדור, ועשרים וארבעה אלף תלמידים ישבו לרגליו!
-
-רבי עקיבא לימד: "ואהבת לרעך כמוך – זה כלל גדול בתורה."`,
-    lesson: 'אף פעם לא מאוחר ללמוד. עם נחישות ואהבה אפשר להגיע לכל מקום.',
-    questions: [
-      {
-        question: 'מה עשה רבי עקיבא לפני שהתחיל ללמוד תורה?',
-        answers: ['לימד ילדים', 'היה רועה צאן', 'בנה בתים', 'מכר בשוק'],
-        correctIndex: 1,
-      },
-      {
-        question: 'מה עורר את רבי עקיבא להתחיל ללמוד?',
-        answers: [
-          'חלום שחלם',
-          'מים שחצבו סלע',
-          'ספר שמצא',
-          'חכם שפגש',
-        ],
-        correctIndex: 1,
-      },
-      {
-        question: 'כמה תלמידים היו לרבי עקיבא?',
-        answers: ['אלף', 'עשרה אלפים', 'עשרים וארבעה אלף', 'מאה'],
-        correctIndex: 2,
-      },
-    ],
-  },
-  {
-    id: 'chofetz-chaim',
-    name: 'החפץ חיים',
-    years: 'תקנ"ח – תרצ"ג (1838–1933)',
-    emoji: '🕊️',
-    color: 'from-green-500 to-teal-600',
-    bgGradient: 'from-green-50 to-teal-100',
-    shortTitle: 'שומר הלשון',
-    story: `רבי ישראל מאיר הכהן מראדין, המכונה "החפץ חיים" על שם ספרו, היה אחד הצדיקים הגדולים של הדור האחרון.
-
-פעם בא אדם לרבי ושאל: "רבי, חטאתי בלשון הרע – מה אעשה?" ענה הרבי: "לך לשוק, קח כר נוצות, ועלה לגג הגבוה ביותר, ושחרר את כל הנוצות לרוח."
-
-עשה האיש כך. חזר אל הרבי ואמר: "עשיתי!" ואמר לו הרבי: "עכשיו לך ואסוף את כל הנוצות!"
-
-"אבל רבי, זה בלתי אפשרי! הרוח פיזרה אותן לכל עבר!"
-
-"כך הם דברי לשון הרע," אמר הרב בעצב. "ברגע שיוצאים מהפה – אי אפשר לאספם בחזרה."
-
-החפץ חיים לימד אותנו לשמור על לשוננו ולדבר רק דברים טובים.`,
-    lesson: 'דברים שאומרים יכולים לפגוע כמו חצים. חשוב לחשוב לפני שמדברים.',
-    questions: [
-      {
-        question: 'על שם מה נקרא "החפץ חיים"?',
-        answers: ['על שם עירו', 'על שם ספרו', 'על שם אביו', 'על שם רבו'],
-        correctIndex: 1,
-      },
-      {
-        question: 'מה ביקש הרב מהאדם לעשות?',
-        answers: [
-          'לבנות בית כנסת',
-          'לצום שלושה ימים',
-          'לשחרר נוצות מהגג',
-          'לתת צדקה',
-        ],
-        correctIndex: 2,
-      },
-      {
-        question: 'מה לימד החפץ חיים בסיפור הנוצות?',
-        answers: [
-          'שהרוח חזקה',
-          'שדברי לשון הרע אי אפשר לקחת בחזרה',
-          'שצריך לנקות את הבית',
-          'שגגות גבוהים מסוכנים',
-        ],
-        correctIndex: 1,
-      },
-    ],
-  },
-  {
-    id: 'rabbi-shimon-bar-yochai',
-    name: 'רבי שמעון בר יוחאי',
-    years: 'המאה השנייה לספירה',
-    emoji: '🔥',
-    color: 'from-red-500 to-rose-600',
-    bgGradient: 'from-red-50 to-rose-100',
-    shortTitle: 'מגלה סודות התורה',
-    story: `רבי שמעון בר יוחאי, המכונה "רשב"י", היה תלמיד של רבי עקיבא ואחד החכמים הגדולים בתלמוד.
-
-כשגזרו הרומאים גזרות קשות על לימוד התורה, ברח רשב"י עם בנו רבי אלעזר ונחבאו במערה שנים-עשר שנה! בזמן הזה למדו תורה כל היום וכל הלילה.
-
-ניס להם נס: צמח עץ חרוב ומזרם מים – כדי שיאכלו וישתו ויוכלו להמשיך ללמוד. בגדיהם הסתירו אותם כשלא למדו, ולבשו אותם רק כשלמדו.
-
-בזכות לימוד התורה הרצוף, גילה רשב"י סודות עמוקים מאוד. ספר הזוהר, הספר הקדוש של הקבלה, מיוחס לו.
-
-בל"ג בעומר, יום פטירתו, ציווה רשב"י לחגוג ולשמוח!`,
-    lesson: 'לימוד התורה הוא אור גדול שמאיר בכל מקום, גם בחשכה.',
-    questions: [
-      {
-        question: 'מי היה רבו של רבי שמעון בר יוחאי?',
-        answers: ['הבעל שם טוב', 'החפץ חיים', 'רבי עקיבא', 'ר׳ יהודה הנשיא'],
-        correctIndex: 2,
-      },
-      {
-        question: 'כמה זמן התחבאו רשב"י ובנו במערה?',
-        answers: ['שנה אחת', 'שלוש שנים', 'שבע שנים', 'שתים-עשרה שנה'],
-        correctIndex: 3,
-      },
-      {
-        question: 'איזה יום חוגגים לזכר רשב"י?',
-        answers: ['ר"ה', 'ל"ג בעומר', 'חנוכה', 'שבועות'],
-        correctIndex: 1,
-      },
-    ],
-  },
-  {
-    id: 'rabbi-meir-baal-hanes',
-    name: "רבי מאיר בעל הנס",
-    years: 'המאה השנייה לספירה',
-    emoji: '🕯️',
-    color: 'from-violet-500 to-purple-600',
-    bgGradient: 'from-violet-50 to-purple-100',
-    shortTitle: 'עושה ניסים',
-    story: `רבי מאיר היה אחד מגדולי תנאי המשנה ותלמיד נאמן של רבי עקיבא. הוא נקרא "בעל הנס" בגלל הניסים שנעשו בזכותו.
-
-אשתו ברוריה הייתה אישה חכמה מאוד. פעם בשבת מתו שני בניהם פתאום. ברוריה לא אמרה לרבי מאיר עד מוצאי שבת, כדי לא לצערו בשבת.
-
-כשסיים שבת, אמרה לו ברוריה: "רבי, שאלה יש לי. מישהו נתן לי לשמור שני תכשיטים יקרים, ועכשיו בא לקחתם חזרה – האם להחזיר?"
-
-"כן, ודאי שמחזירים פיקדון!" ענה הרב.
-
-לקחה אותו ברוריה אל חדר שני ילדיהם ואמרה: "ה׳ נתן וה׳ לקח – יהי שם ה׳ מבורך."
-
-רבי מאיר קיבל את דין השמיים בגבורה ובאמונה שלמה.`,
-    lesson: 'אפשר לקבל גם דברים קשים בחיים עם אמונה ורוגע.',
-    questions: [
-      {
-        question: 'מדוע נקרא "בעל הנס"?',
-        answers: [
-          'כי גר בנס',
-          'כי ניסים נעשו בזכותו',
-          'כי אהב לספר סיפורים',
-          'כי היה מלמד בניסן',
-        ],
-        correctIndex: 1,
-      },
-      {
-        question: 'מה עשתה ברוריה עם ידיעת מות ילדיה?',
-        answers: [
-          'שלחה מיד שליח לבעלה',
-          'בכתה כל השבת',
-          'חיכתה עד מוצאי שבת',
-          'ברחה מהבית',
-        ],
-        correctIndex: 2,
-      },
-      {
-        question: 'מה הייתה תשובת ברוריה לבעלה?',
-        answers: [
-          '"אנחנו עניים"',
-          '"ה׳ נתן וה׳ לקח – יהי שם ה׳ מבורך"',
-          '"הייתה תאונה"',
-          '"הם היו חולים"',
-        ],
-        correctIndex: 1,
-      },
-    ],
-  },
-  {
-    id: 'the-alter-of-kelm',
-    name: "הרב מנחם מנדל מוויטבסק",
-    years: 'תק"ל – תקנ"ח (1730–1788)',
-    emoji: '🌿',
-    color: 'from-lime-500 to-green-600',
-    bgGradient: 'from-lime-50 to-green-100',
-    shortTitle: 'אהבת ארץ ישראל',
-    story: `רבי מנחם מנדל מוויטבסק היה תלמיד גדול של המגיד ממזריץ', ממשיכי הבעל שם טוב.
-
-מתוך אהבה עזה לארץ ישראל, עלה רבי מנחם מנדל לארץ ישראל עם שלוש מאות חסידים בשנת תקל"ז. זה היה מסע קשה מאוד – חודשים של הפלגה בים ומסע יבשה.
-
-הם נקלטו בצפת ואחר כך עברו לטבריה. החיים היו קשים מאוד – עוני, מחלות, וקשיים רבים. חסידיו שאלוהו פעמים רבות: "רבי, למה לא נחזור?"
-
-ענה רבם: "אני ישן על אדמת ישראל, ואפילו שעה אחת כאן שווה יותר מכל שנות החיים שם!"
-
-גם מרחוק המשיך לכתוב אגרות עידוד לחסידיו שנשארו בגלות, ועודדם לשמור אמונה ולחכות לגאולה.`,
-    lesson: 'ארץ ישראל היא ארץ מיוחדת, ואהבתה שווה כל מחיר.',
-    questions: [
-      {
-        question: 'מה עשה רבי מנחם מנדל מתוך אהבה לארץ ישראל?',
-        answers: [
-          'כתב ספר על הארץ',
-          'שלח כסף לעניים',
-          'עלה לארץ ישראל עם שלוש מאות חסידים',
-          'לימד ילדים על הארץ',
-        ],
-        correctIndex: 2,
-      },
-      {
-        question: 'באיזו עיר התיישבו החסידים בתחילה?',
-        answers: ['ירושלים', 'חברון', 'צפת', 'באר שבע'],
-        correctIndex: 2,
-      },
-      {
-        question: 'מה אמר הרב כשרצו לחזור?',
-        answers: [
-          '"אז נחזור"',
-          '"שעה אחת בארץ ישראל שווה יותר מכל שנות החיים שם"',
-          '"זה קשה מדי"',
-          '"צדקתם, נחזור"',
-        ],
-        correctIndex: 1,
-      },
-    ],
-  },
+  BAAL_SHEM_TOV_STORY,
+  RABBI_AKIVA_STORY,
+  CHOFETZ_CHAIM_STORY,
+  RABBI_SHIMON_BAR_YOCHAI_STORY,
+  RABBI_MEIR_BAAL_HANES_STORY,
+  MENACHEM_MENDEL_MI_VITEBSK_STORY,
 ];
